+++ src/types/types.ts
@@ -111,3 +111,16 @@ export interface SectionProps {
   className?: string;
   contentClassName?: string;
 }
+
+export interface Product {
+  slug: string;
+  name: string;
+  category: string;
+  description: string;
+  owner: string;
+  slackChannel?: string;
+  docsUrl?: string;
+  externalUrl?: string;
+  status: StatusTagValue;
+  tags: string[];
+}
